refactor(ui): MetricsTreemap - extract display size resolvers & tooltip state hook into utils
