+++ packages/ui/src/components/metrics-treemap/metrics-treemap.constants.ts
@@ -50,3 +50,14 @@ export interface Tree {
 export type TreeNode = Tree | TreeLeaf;
 
 export type TreeNodeChildren = Array<TreeNode>;
+
+export type TileGroupSizeDisplay = 'minimal' | 'small' | 'default';
+
+export type TileSizeDisplay = 'minimal' | 'small' | 'default';
+
+export const SQUARIFY_RATIO = 1.66;
+export const PADDING_OUTER = 1;
+export const PADDING_INNER = 2;
+export const NESTED_PADDING = 4;
+export const NESTED_PADDING_TOP = 16 + NESTED_PADDING * 2;
+export const NESTED_PADDING_LEFT = 8;
