+++ packages/ui/src/components/metrics-treemap/metrics-treemap.tsx
@@ -9,10 +9,10 @@ import React, {
   useMemo,
   useRef,
 } from 'react';
-import { useDebounce, useHoverDirty, useMeasure, useMouseHovered } from 'react-use';
+import { useHoverDirty, useMeasure } from 'react-use';
 import cx from 'classnames';
 import { HierarchyRectangularNode, hierarchy, treemap, treemapSquarify } from 'd3';
-import { Tooltip, TooltipArrow, TooltipAnchor, useTooltipState } from 'ariakit/tooltip';
+import { Tooltip, TooltipArrow, TooltipAnchor } from 'ariakit/tooltip';
 import {
   type ReportMetricRow,
   type MetricRunInfo,
@@ -25,113 +25,26 @@ import { Stack } from '../../layout/stack';
 import { FileName } from '../../ui/file-name';
 import { Delta } from '../delta';
 import { RunInfo } from '../run-info';
-import type { TreeLeaf, TreeNode, Tree } from './metrics-treemap.constants';
+import {
+  type TreeLeaf,
+  type TreeNode,
+  type Tree,
+  SQUARIFY_RATIO,
+  PADDING_OUTER,
+  PADDING_INNER,
+  NESTED_PADDING,
+  NESTED_PADDING_TOP,
+  NESTED_PADDING_LEFT,
+  TileSizeDisplay,
+} from './metrics-treemap.constants';
 import * as I18N from './metrics-treemap.i18n';
 import css from './metrics-treemap.module.css';
-import { resolveGroupDeltaType } from './metrics-treemap.utils';
-
-const SQUARIFY_RATIO = 1.66;
-const PADDING_OUTER = 1;
-const PADDING_INNER = 2;
-const NESTED_PADDING = 4;
-const NESTED_PADDING_TOP = 16 + NESTED_PADDING * 2;
-const NESTED_PADDING_LEFT = 8;
-
-/**
- * Resolve the tile's size using predefined values to avoid
- * computing the size of the content for every tile
- */
-const PADDING_TOP = 8;
-const PADDING_BOTTOM = 8;
-const PADDING_LEFT = 8;
-const PADDING_RIGHT = 8;
-const LINE_HEIGHT = 16;
-const LINE_HEIGHT_SMALL = 13.3;
-
-type TileGroupSizeDisplay = 'minimal' | 'small' | 'default';
-
-function resolveTileGroupSizeDisplay(width: number, height: number): TileGroupSizeDisplay {
-  if (height < NESTED_PADDING_TOP) {
-    return 'minimal';
-  }
-
-  if (width < 24) {
-    return 'minimal';
-  }
-
-  if (height < NESTED_PADDING_TOP + 8) {
-    return 'small';
-  }
-
-  return 'default';
-}
-
-type TileSizeDisplay = 'minimal' | 'small' | 'default';
-
-function resolveTileSizeDisplay(width: number, height: number): TileSizeDisplay {
-  if (height < PADDING_TOP + PADDING_BOTTOM + LINE_HEIGHT_SMALL) {
-    return 'minimal';
-  }
-
-  if (width < PADDING_LEFT + PADDING_RIGHT + 42) {
-    return 'minimal';
-  }
-
-  if (height < PADDING_TOP + PADDING_TOP + LINE_HEIGHT * 2) {
-    return 'small';
-  }
-
-  if (width < PADDING_LEFT + PADDING_RIGHT + 96) {
-    return 'small';
-  }
-
-  return 'default';
-}
-
-interface UseTooltipStateWithMouseFollowOptions {
-  /**
-   * React ref to parent div
-   */
-  parentRef: RefObject<Element>;
-  /**
-   * Tooltip gutter
-   */
-  gutter?: number;
-  /**
-   * Tooltip timeout
-   */
-  timeout?: number;
-}
-
-function useTooltipStateWithMouseFollow(options: UseTooltipStateWithMouseFollowOptions) {
-  const { parentRef, gutter = 16, timeout = 180 } = options;
-
-  const pointer = useMouseHovered(parentRef, { whenHovered: true });
-
-  // Return custom rect based on the mouse position
-  const getAnchorRect = useCallback(() => {
-    // Skip custom component rect area if the pointer position is empty
-    if (pointer.docX === 0 && pointer.docY === 0 && pointer.elW === 0 && pointer.elH === 0) {
-      return null;
-    }
-
-    const newRect = {
-      x: pointer.docX - (document?.defaultView?.scrollX ?? 0),
-      y: pointer.docY - (document?.defaultView?.scrollY ?? 0),
-      w: 1,
-      h: 1,
-    };
-
-    return newRect;
-  }, [pointer.docX, pointer.docY]);
-
-  const tooltipState = useTooltipState({ gutter, getAnchorRect, timeout });
-
-  // Update tooltip position when poiner values are changing
-  useDebounce(tooltipState.render, 5, [pointer.docX, pointer.docY]);
-
-  return tooltipState;
-}
+import {
+  resolveGroupDeltaType,
+  resolveTileSizeDisplay,
+  resolveTileGroupSizeDisplay,
+  useTooltipStateWithMouseFollow,
+} from './metrics-treemap.utils';
 
 interface TileTooltipContentProps {
   item: ReportMetricRow;
